Notification 테스트 코드 추가 및 수정, swagger 정리

+++ src/main/java/com/project/devidea/modules/notification/NotificationController.java
@@ -1,11 +1,13 @@
 package com.project.devidea.modules.notification;
 
+import com.project.devidea.infra.config.security.LoginUser;
 import com.project.devidea.infra.error.exception.EntityNotFoundException;
 import com.project.devidea.modules.account.Account;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,32 +21,31 @@ public class NotificationController {
     private final NotificationRepository notificationRepository;
 
     @GetMapping("/new")
-    public ResponseNotifications<List<Notification>> getNewNotifications(Long id) {
-        // 시큐리티 적용전 임시
-        Account account = new Account().builder().id(id).build();
+    public ResponseNotifications<List<Notification>> getNewNotifications(@AuthenticationPrincipal LoginUser loginUser) {
+        Account account = loginUser.getAccount();
         List<Notification> notifications = notificationRepository.findByAccountAndCheckedOrderByCreatedDateTimeDesc(account, false);
         int countOldNotifications = notificationRepository.countByAccountAndChecked(account, true);
         return new ResponseNotifications(notifications.size(), countOldNotifications, notifications);
     }
 
     @GetMapping("/old")
-    public ResponseNotifications<List<Notification>> getOldNotifications(Long id) {
-        // 시큐리티 적용전 임시
-        Account account = new Account().builder().id(id).build();
+    public ResponseNotifications<List<Notification>> getOldNotifications(@AuthenticationPrincipal LoginUser loginUser) {
+        Account account = loginUser.getAccount();
         List<Notification> notifications = notificationRepository.findByAccountAndCheckedOrderByCreatedDateTimeDesc(account, true);
         int countNewNotifications = notificationRepository.countByAccountAndChecked(account, false);
         return new ResponseNotifications(notifications.size(), countNewNotifications, notifications);
     }
 
     @PostMapping("/delete")
-    public void deleteAllOldNotifications(Long id) {
-        // 시큐리티 적용전 임시
-        Account account = new Account().builder().id(id).build();
-        notificationRepository.deleteByAccountAndChecked(account, true);
+    public void deleteAllOldNotifications(@AuthenticationPrincipal LoginUser loginUser) {
+        notificationRepository.deleteByAccountAndChecked(loginUser.getAccount(), true);
     }
 
     @PostMapping("/{id}/delete")
     public void deleteNotification(@PathVariable("id") Long id) {
+        Optional<Notification> findNotification = notificationRepository.findById(id);
+        if(!findNotification.isPresent())
+            throw new EntityNotFoundException("notification");
         notificationRepository.deleteById(id);
     }
